Add a class for git-annex matching options
I guess I won't need most of it soon, but anyway. Since git-annex find
is eventually going to take a list of option strings, making this into
something like a proper and-or tree doesn't feel like it will be worth
it. Subclassing list lets me just pass this object itself as part of the
command line, and subclassing UserList lets me combine it with other
lists and still end up with this class (instead of the parent, list).