Add a function to annexed file tree that runs git-annex find
It just returns a list of file paths relative to the tree we run it
on. The find --json call also returns backend, bytesize, hashdirlower,
hashdirmixed, humansize, key, keyname, mtime (in addition to file); but
I think accessing these properties should be done through AnnexedFile
(but that's still not implemented).